Tyuiu.GalimovaAS.Sprint2.Task6.V4.Test/DataServiceTest.cs
using Tyuiu.GalimovaAS.Sprint2.Task6.V4.Lib;
namespace Tyuiu.GalimovaAS.Sprint2.Task6.V4.Test
{
    [TestClass]
    public class DataServiceTest
    {
        [TestMethod]
        public void TestMethod1()
        {
            DataService ds = new DataService();

            Assert.AreEqual("пики", ds.FindCardSuit(1));
            Assert.AreEqual("трефы", ds.FindCardSuit(2));
            Assert.AreEqual("бубны", ds.FindCardSuit(3));
            Assert.AreEqual("черви", ds.FindCardSuit(4));
            Assert.ThrowsException<ArgumentException>(() =>
            {
                ds.FindCardSuit(5);
            });
            Assert.ThrowsException<ArgumentException>(() =>
            {
                ds.FindCardSuit(-1);
            });
        }
    }
}